requirements-test.txt
pytest>=7.4.0
PyYAML>=6.0.1
